move parser tests into own files per parse func
