+++ pkg/parser/defaultvalue_parser_test.go
@@ -1,0 +1,18 @@
+package parser
+
+import (
+	"github.com/jensneuse/graphql-go-tools/pkg/document"
+	"testing"
+)
+
+func TestParser_parseDefaultValue(t *testing.T) {
+	t.Run("integer", func(t *testing.T) {
+		run("= 2", mustParseDefaultValue(document.ValueTypeInt))
+	})
+	t.Run("bool", func(t *testing.T) {
+		run("= true", mustParseDefaultValue(document.ValueTypeBoolean))
+	})
+	t.Run("invalid", func(t *testing.T) {
+		run("true", mustPanic(mustParseDefaultValue(document.ValueTypeBoolean)))
+	})
+}
